pyproject.toml
[tool.poetry]
name = "agentwiki"
version = "0.1.0"
description = ""
authors = ["Pilar Madariaga <prng@gft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = ">=0.1.9"
boto3 = "^1.34.50"
boto3-type-annotations = "^0.3.1"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"
langchain-community = "^0.0.33"
wikipedia = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
